requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==25.1.0
    # via
    #   outcome
    #   trio
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
    #   selenium
cffi==1.17.1
    # via trio
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
defusedxml==0.7.1
    # via youtube-transcript-api
distro==1.9.0
    # via openai
flask==3.1.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
jiter==0.8.2
    # via openai
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
openai==1.63.0
    # via -r requirements.in
outcome==1.3.0.post0
    # via trio
packaging==24.2
    # via gunicorn
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pysocks==1.7.1
    # via urllib3
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via youtube-transcript-api
selenium==4.28.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
    #   trio
sortedcontainers==2.4.0
    # via trio
tqdm==4.67.1
    # via openai
trio==0.29.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   selenium
urllib3[socks]==2.3.0
    # via
    #   requests
    #   selenium
websocket-client==1.8.0
    # via selenium
werkzeug==3.1.3
    # via flask
wsproto==1.2.0
    # via trio-websocket
youtube-transcript-api==0.6.3
    # via -r requirements.in
